requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-crispy-forms==1.11.2
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
